+++ lib/interfaces/rmq-options.interface.ts
@@ -2,6 +2,7 @@ import { RMQPipeClass } from '../classes/rmq-pipe.class';
 import { RMQIntercepterClass } from '../classes/rmq-intercepter.class';
 import { RMQErrorHandler } from '../classes/rmq-error-handler.class';
 import { LoggerService } from '@nestjs/common';
+import { ModuleMetadata } from '@nestjs/common/interfaces';
 
 export interface IRMQServiceOptions {
 	exchangeName: string;
@@ -15,6 +16,7 @@ export interface IRMQServiceOptions {
 	isQueueDurable?: boolean;
 	isExchangeDurable?: boolean;
 	reconnectTimeInSeconds?: number;
+	heartbeatIntervalInSeconds?: number;
 	messagesTimeout?: number;
 	logMessages?: boolean;
 	logger?: LoggerService;
@@ -28,4 +30,9 @@ export interface IRMQConnection {
 	login: string;
 	password: string;
 	host: string;
+}
+
+export interface IRMQServiceAsyncOptions extends Pick<ModuleMetadata, 'imports'> {
+	useFactory?: (...args: any[]) => Promise<IRMQServiceOptions> | IRMQServiceOptions;
+	inject?: any[];
 }